+++ app/Http/Controllers/LikeController.php
@@ -14,11 +14,13 @@ class LikeController extends Controller {
 
     public function destroy( $id, ApiResponse $response ) {
         $like = Like::where( 'id', '=', $id )
-                    ->where( 'author', '=', auth()->user()->id )
                     ->first();
 
         if ( is_null( $like ) )
             abort( 404 );
+
+        if ( ! $this->canTouch( $like ) )
+            abort( 403 );
 
         try {
             $like->delete();
@@ -37,6 +39,16 @@ class LikeController extends Controller {
         }
 
         return $response->success( 'Like deleted successfully' );
+    }
+
+    private function canTouch( $like ) {
+        if ( auth()->user()->group_id <= 2 )
+            return true;
+
+        if ( $like->author == auth()->user()->id )
+            return true;
+
+        return false;
     }
     
 }
